Rename match maker to match repo, it is not a service
