+++ app/com/codelab27/cards9/services/loader/Cards9ApplicationLoader.scala
@@ -2,7 +2,7 @@ package com.codelab27.cards9.services.loader
 
 import com.codelab27.cards9.controllers.MatchMakerController
 import com.codelab27.cards9.routes.GameRouter
-import com.codelab27.cards9.services.matchmaking.mem.MatchMakerInMemoryInterpreter
+import com.codelab27.cards9.repos.matches.mem.MatchRepositoryInMemoryInterpreter
 
 import play.api.ApplicationLoader.Context
 import play.api.{ApplicationLoader, BuiltInComponentsFromContext, NoHttpFiltersComponents}
@@ -17,8 +17,8 @@ class Cards9Components(context: Context) extends BuiltInComponentsFromContext(co
 
   import com.codelab27.cards9.utils.DefaultCatsInstances._
 
-  lazy val matchMaker = MatchMakerInMemoryInterpreter
-  lazy val matchMakerController = new MatchMakerController(controllerComponents, matchMaker)
+  lazy val matchRepo = MatchRepositoryInMemoryInterpreter
+  lazy val matchMakerController = new MatchMakerController(controllerComponents, matchRepo)
 
   lazy val router = new GameRouter(
     matchMakerController
